made calculator to compute the number of possible combinations when selecting a sample of a population
equation:
n!/k!(n-k)!
where
n = the population size
k the size of the sample you are selecting
for example say you have 5 letters (a, b, c, d, e), and you were going to randomly select 2 of them. there are 10 different possible combinations that you will get when the order they are selected doesn't matter.
the 10 possible combinations would be:
a and b
a and c
a and d
a and e
b and c
b and d
b and e
c and d
c and e
d and e